list post fixed

+++ src/Controller/Admin/PostController.php
@@ -77,9 +77,15 @@ class PostController extends BackendController
     public function deletePost($id): Response
     {
         $post = $this->postService->getPost($id);
+
+        if ($post === null) {
+            $this->addFlash('errors', 'Post not found !');
+            return $this->redirectToRoute('list_posts');
+        }
+        else{
         $this->postService->deletePost($post);
         $this->addFlash('success', 'Post has been deleted successfully !');
-
+        }
         return $this->redirectToRoute('list_posts');
     }
 }
